[Mod] move brutal force optimization function to vnpy.trader.optimize
