+++ vnpy/trader/optimize.py
@@ -1,0 +1,78 @@
+from typing import Dict, List, Callable, Tuple
+from itertools import product
+from concurrent.futures import ProcessPoolExecutor
+
+
+class OptimizationSetting:
+    """
+    Setting for runnning optimization.
+    """
+
+    def __init__(self):
+        """"""
+        self.params = {}
+        self.target_name = ""
+
+    def add_parameter(
+        self, name: str, start: float, end: float = None, step: float = None
+    ):
+        """"""
+        if not end and not step:
+            self.params[name] = [start]
+            return
+
+        if start >= end:
+            print("参数优化起始点必须小于终止点")
+            return
+
+        if step <= 0:
+            print("参数优化步进必须大于0")
+            return
+
+        value = start
+        value_list = []
+
+        while value <= end:
+            value_list.append(value)
+            value += step
+
+        self.params[name] = value_list
+
+    def set_target(self, target_name: str):
+        """"""
+        self.target_name = target_name
+
+    def generate_setting(self):
+        """"""
+        keys = self.params.keys()
+        values = self.params.values()
+        products = list(product(*values))
+
+        settings = []
+        for p in products:
+            setting = dict(zip(keys, p))
+            settings.append(setting)
+
+        return settings
+
+    def generate_setting_ga(self):
+        """"""
+        settings_ga = []
+        settings = self.generate_setting()
+        for d in settings:
+            param = [tuple(i) for i in d.items()]
+            settings_ga.append(param)
+        return settings_ga
+
+
+def run_bf_optimization(
+    optimization_func: Callable,
+    optimization_setting: OptimizationSetting,
+    max_workers: int = None
+) -> List[Tuple]:
+    """Run brutal force optimization"""
+    with ProcessPoolExecutor(max_workers) as executor:
+        settings: List[Dict] = optimization_setting.generate_setting()
+        results: List[Tuple] = list(executor.map(optimization_func, settings))
+        results.sort(reverse=True, key=lambda result: result[1])
+        return results
